NP-34646 Minor changes
